Use DebitCollectorService

+++ classes/DebtCollector.php
@@ -6,6 +6,6 @@ class DebtCollector implements \Xi\Php\Interfaces\DebtCollectorInterface
 {
     public function collect(float $amount): float
     {
-        return mt_rand($amount * 0.5, $amount);
+        return mt_rand((int) ($amount * 0.5), (int) $amount);
     }
 }
